removed old chat code
