+++ app/build.gradle
@@ -35,8 +35,8 @@ android {
         applicationId "elliott.software.clicker"
         minSdk 24
         targetSdk 34
-        versionCode 13
-        versionName "1.12"
+        versionCode 14
+        versionName "1.13"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables {
